Moved Pagination.get_items to PaginationInternal.get_pagination_items so we can get at it in the unit tests.
